+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Check for mandatory command line arguments (ER #14)
 - Exit program immediately if no files were found (ER #17)
 - More debug statements for file reading operations (ER #15)
+- Color coded output (can be turned off via $NO_COLOR) (ER #20)
 
 ### Fixed
 - Pass correct exit return codes (bug #18)